requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
distro==1.9.0
dotenv-python==0.0.1
etelemetry==0.3.1
filelock==3.14.0
fitz==0.0.1.dev2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
isodate==0.6.1
looseversion==1.3.0
lxml==5.2.2
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.33.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.3.0
prov==2.0.1
pydantic==2.7.3
pydantic_core==2.18.4
pydot==2.0.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
